The default value of 'public' doesn't make sense.
Given that the call gets the public repositories of the specified user, it
doesn't make sense that the default value of type would be 'public' which
isn't even listed as an option. Instead it makes more sense that it would
default to 'all' since you could then narrow down the search to 'owner' or
'member'.