Facebook API redirects to Facebook Login guide
URL needs to be updated along with title so as to not land the
user on a page they think is wrongly linked/called. Pointing
to their technical guides might also be warmer than a straight
drop into the API specs.

Also spruces up some of those GitHub links in OAuth since all
are on the SSL flavour.

That Net::Github CPAN link also moved house.