Remove misleading parameters in notifications API
The API allows you to mark notifications as read. To do so, you can
specify any one of the following parameters and values.

  - "unread" : false
  - "read" : true
  - none of the above

In all cases, the result is the same: it marks the notifications as
read. Therefore, passing neither is the simplest approach, and there is
no reason to describe these parameters.