Announce plans to fix `type` value for submodules
Prior to 2013-04-30, you could not GET a submodule via the Repository
Contents API. (You could see the submodule listed when you GET the
directory that the submodule resides it, but you could not GET just the
submodule.) On 2013-04-30, we added the ability to GET a submodule.
(See 81c7e58.) When you GET a submodule, the "type" is returned as
"submodule".

Prior to 2013-04-30, when you listed the contents of a directory using
the Repository Contents API, the output included submodules, but it
returned the "type" as "file". The type *should* be "submodule". We are
leaving this behavior as is for v3, so as not to break any client code
that might depend on this behavior. In the next major version of the
API, the type will be returned as "submodule".