Use `stargazers` (instead of `watchers`) to sort forks
In API v3, for backwards compatibility purposes, watchers and stargazers are the
same thing. [1]

API v3 now supports `stargazers` as an option for sorting the list of forks for
a repo. API v3 will continue to support `watchers` as an option for sorting,
but we're no longer advertising that support. It does the same thing as
`stargazers`. Any new API client code should use `stargazers` instead of
`watchers`.

[1]: http://developer.github.com/changes/2012-9-5-watcher-api/