Add `stargazers_count` to summary repository representation
In API v3, for backwards compatibility purposes, watchers and stargazers are the
same thing. [1] So the watchers_count is **really** the number of stargazers.
Let's add a stargazers_count attribute as one small step toward helping API
consumers figure out how to get the number of stargazers for a repo.

[1]: http://developer.github.com/changes/2012-9-5-watcher-api/