Remove downloads methods from deprecations list.
The Downloads API was deprecated December 11, 2012. It may be
disabled before the next major version of the API.

For this reason it makes more sense to exclude these methods from
the Deprecations list, which lists functionality which will be
removed in the next major version of the API.

The copy of the alert on the Downloads API page has been modified
to be:

The Downloads API (described below) was deprecated on December 11,
2012. It will be removed at a future date. We recommend using
Releases instead.