Fix example link headers for Events API
By default, the docs show a "next" and "last" link in the link header.
However, the Events API only provides "first", "prev", and "next" links
in the link header. Therefore, when you get the first page, the link
header only includes a "next" link. Update the example response to
reflect that behavior.