Fix overall heirarchy of hooks page
I've seen this for a while but didn't realize how bad it was
until I started adding to the page. I've put this in a separate
commit so we can just revert it if you don't want to use it. I
think this makes the TOC much more useful. @jdennes @izuzak
@github/docs