Update gist files hash to consistently key by filename
The responses returned by the API for Gists always includes a hash
for files that contains a hash for each file in the gist keyed by the
file's name, but this was only reflected in the example response for
listing gists. This commit updates the example responses for getting a
single gist, creating a gist, and editing a gist to correct the files
hash.