Deprecate plan["collaborators"] in user JSON
Personal plans and organization plans used to allow a finite number of
collaborators. The API used to return the allowed number of
collaborators for each plan in the `collaborators` attribute:

{
  "login": "jasonrudolph",
  ...
  "plan": {
    "name": "medium",
    "space": 2516582,
    "collaborators": 10,
    "private_repos": 20
  }
}

Now, these plans allow unlimited collaborators. As a result, this
attribute is no longer meaningful. We will remove it in the next major
version of the API. In the meantime, for backwards compatibility
purposes, the API continues to return this attribute and continues to
provide an integer value (0) for the attribute.

{
  "login": "jasonrudolph",
  ...
  "plan": {
    "name": "medium",
    "space": 2516582,
    "collaborators": 0,
    "private_repos": 20
  }
}