Update v3.md
It seems to be better to mention RFC 3339 instead of ISO 8601. RFC 3339 is far simpler, clear and easier to use unlike complex and vague ISO 8601. Then it can decrease a lot of validation concerns of API users. Anyway this is based on my guess, so maybe wrong.

Though I am not a Github employee, I am sending this PR because the public site is saying it's encouraged.