Broken If-Modified-Since + no 'since' = a bit clunky
*This would be better raised as an issue but you have them turned off.  The patch is not to be taken entirely seriously.*

The `*/events` endpoints:

 * return a `Last-Modified` header, but it's inaccurate and doesn't change when ETag does (you'll easily find events streams containing events paradoxically newer than the `Last-Modified` header)
 * do not support `since=<timestamp>`

`Last-Modified` should be made accurate, removed or should carry a warning, to minimise confusion.

Current behaviour means that a caller wanting to efficiently process only events since they last checked must store both the ETag (to avoid unconditional requests) and the timestamp (to manually discard events older than that) and still needs to request more events than they care about.  It might be neater for callers and lighter on GitHub's bandwidth to make `Last-Modified` accurate and amenable to conditional requests, and/or to support `since` so that callers need only grab the events since their last check.  That way, a single timestamp would be enough state.