Clarify that draft releases can't be looked up by tag name
The previous documentation seemed to hint that you can look up a draft release by it's tag name (which is not yet a real git tag). This is not true, since the API will return 404 if the git tag name doesn't exist. The draft release will have to be looked up from the list of all releases.